// admin/src/api/generated/models/ApplicationEnvPlaintextResponse.ts
export interface ApplicationEnvPlaintextResponse {
    /**
     *
     * @type {string}
     * @memberof ApplicationEnvPlaintextResponse
     */
    applicationId: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationEnvPlaintextResponse
     */
    content: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationEnvPlaintextResponse
     */
    digest: string;
    /**
     *
     * @type {number}
     * @memberof ApplicationEnvPlaintextResponse
     */
    envVersion: number;
    /**
     *
     * @type {string}
     * @memberof ApplicationEnvPlaintextResponse
     */
    fileName: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationEnvPlaintextResponse
     */
    format: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationEnvPlaintextResponse
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationEnvPlaintextResponse
     */
    module: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationEnvPlaintextResponse
     */
    updatedAt: string;
    /**
     *
     * @type {number}
     * @memberof ApplicationEnvPlaintextResponse
     */
    version: number;
}