// admin/src/api/generated/models/EnvGrantAction.ts
/* tslint:disable */
/* eslint-disable */
/**
 * deploy-go-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 *
 * @export
 */
export const EnvGrantAction = {
    ReadWrite: 'read_write',
    Delete: 'delete'
} as const;
export type EnvGrantAction = typeof EnvGrantAction[keyof typeof EnvGrantAction];


export function instanceOfEnvGrantAction(value: any): boolean {
    for (const key in EnvGrantAction) {
        if (Object.prototype.hasOwnProperty.call(EnvGrantAction, key)) {
            if (EnvGrantAction[key as keyof typeof EnvGrantAction] === value) {
                return true;
            }
        }
    }
    return false;
}

export function EnvGrantActionFromJSON(json: any): EnvGrantAction {
    return EnvGrantActionFromJSONTyped(json, false);
}

export function EnvGrantActionFromJSONTyped(json: any, ignoreDiscriminator: boolean): EnvGrantAction {
    return json as EnvGrantAction;
}

export function EnvGrantActionToJSON(value?: EnvGrantAction | null): any {
    return value as any;
}

export function EnvGrantActionToJSONTyped(value: any, ignoreDiscriminator: boolean): EnvGrantAction {
    return value as EnvGrantAction;
}